+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -275,6 +275,7 @@ class SkyHanniMod {
         loadModule(BingoCardTips())
         loadModule(GardenVisitorDropStatistics)
         loadModule(SackDisplay())
+        loadModule(GardenStartLocation)
 
         init()